+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 2.6.0
 
+  - deprecated mak run (use run:mame instead)
+
 ## 2.5.1
 
   - rename few static functions